changes based on switch statement review
I did not change much, even though the answers were better in some ways. I decided to keep things simple for long-term use.

I could have made things even simpler if I had switched on a literal value for age, instead of a variable's set value.

It was a good call on accommodating for other cases, such as for humans less than a year old and I learned of a way to implement an open-ended upper limit for ages greater than 61.